refactor into ./serviceregistration/consul
