+++ command/commands.go
@@ -64,6 +64,7 @@ import (
 	physInmem "github.com/hashicorp/vault/sdk/physical/inmem"
 
 	sr "github.com/hashicorp/vault/serviceregistration"
+	csr "github.com/hashicorp/vault/serviceregistration/consul"
 )
 
 const (
@@ -159,7 +160,7 @@ var (
 	}
 
 	serviceRegistrations = map[string]sr.Factory{
-		"consul": physConsul.NewConsulServiceRegistration,
+		"consul": csr.NewConsulServiceRegistration,
 	}
 )
